style: :lipstick: change date picker styles
Some style changes and a function that if a day outside of month is selected, the shown month also changes. Removed stale code & commits. Add new info comits. Note: There was a problem with date-picker component height as it sometimes displayed an additional week in some months, but it got fixed by itself.