feat: :sparkles: replace content textarea with rich text editor
Textarea on the home page is replaced by rich text editor using lexical. The function to save an entry was separated into two parts - setEntryTitle and setEntryContent - which solved the `last change not saved` issue. Global content state was removed.

BREAKING CHANGE: textarea content is not compatible with editorState saved in the content field.
Fixes #36